mount: separate mount error wrapper
This is only used for !windows, so move this to a separate
source file to avoid the following windows linter warning:

> ##[error]mount.go:14:6: type `mountError` is unused (unused)

The code is not changed.
